requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
opencv-python
python-multipart